backend/requirements.txt
# Web Framework
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
click==8.1.8
fastapi==0.115.6
h11==0.16.0
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0

# WebSocket / Socket.IO
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
sniffio==1.3.1
wsproto==1.2.0

# Database
mysql-connector-python==9.1.0
SQLAlchemy>=2.0
alembic>=1.10
aiosqlite>=0.19
sqlmodel>=0.0.8
aiomysql>=0.1

# Env
python-dotenv>=1.0.0

# BLE
bless>=0.2.5
aioconsole>=0.7.0
bleak>=0.20.0

# Raspberry Pi GPIO (uncomment when on RPi)
#lgpio==0.2.2.0
#rpi-lgpio==0.6